created cpp file for client and cleaned main file
